+++ tests/sketches/ledsignal/ledsignal.h
@@ -4,9 +4,11 @@
 // If something went wrong, it will flash every 0.2 seconds for 0.2 seconds
 // For the debugger, two routines are provided for testing the outcome.
 #include <util/delay.h>
+// Always blink SCK
 #ifdef SCK
   #define LED0 SCK
 #endif
+// Additionally use the "standard" LED
 #if defined(__AVR_ATtiny13__) || defined(__AVR_ATtiny25__) || defined(__AVR_ATtiny45__) || \
   defined(__AVR_ATtiny85__)
 #define LED1 4
@@ -20,35 +22,47 @@
 #endif
 
 
-void OKBlink() {
-  _delay_ms(1500);
-}
-
-void FailBlink() {
-  _delay_ms(200);
-}
-
-void blink_exit (boolean OK) {
-#ifdef LED0
-  pinMode(LED0, OUTPUT);
-#endif
-#ifdef LED1
-  pinMode(LED1, OUTPUT);
-#endif
-  while (1) {
+void LedOn() {
 #ifdef LED0
     digitalWrite(LED0, HIGH);
 #endif
 #ifdef LED1
     digitalWrite(LED1, HIGH);
 #endif
-    _delay_ms(200);
+}
+
+void LedOff() {
 #ifdef LED0
     digitalWrite(LED0, LOW);
 #endif
 #ifdef LED1
     digitalWrite(LED1, LOW);
 #endif
+}
+
+void LedInit() {
+#ifdef LED0
+  pinMode(LED0, OUTPUT);
+#endif
+#ifdef LED1
+  pinMode(LED1, OUTPUT);
+#endif
+}
+
+void OKBlink() {
+  delay(1500);
+}
+
+void FailBlink() {
+  delay(200);
+}
+
+void blink_exit (boolean OK) {
+  LedInit();
+  while (1) {
+    LedOn();
+    delay(200);
+    LedOff();
     if (OK) OKBlink();
     else FailBlink();
   }